requirements.txt
fastapi
APScheduler
newspaper3k
lxml[html_clean]
feedparser
python-dotenv
langchain
langchain-community
huggingface-hub
